Resolved merge conflict and adapted with the new layout.
