+++ src/App.css
@@ -1,100 +1,33 @@
 .App {
-    text-align: center;
+  text-align: center;
 }
 
 .App-logo {
-    height: 40vmin;
-      pointer-events: none;
+  height: 40vmin;
+    pointer-events: none;
 }
 
 @media (prefers-reduced-motion: no-preference) {
-    .App-logo {
-        animation: App-logo-spin infinite 20s linear;
-      }
+  .App-logo {
+      animation: App-logo-spin infinite 20s linear;
+    }
 }
 
 .App-header {
-    background-color: #282c34;
-      min-height: 100vh;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      justify-content: center;
-      font-size: calc(10px + 2vmin);
-      color: white;
-    position: relative;
-    }
+  background-color: white;
+    min-height: 100vh;
+  position: relative;
+  }
 
-    .user-info {
-      position: absolute;
-      top: 20px;
-      right: 20px;
-      display: flex;
-      align-items: center;
-      cursor: pointer;
-    }
 
-    .user-info img {
-      width: 50px;
-      height: 50px;
-      border-radius: 50%;
-      margin-right: 10px;
-    }
-
-    .user-info span {
-      font-size: 1.2rem;
-      color: white;
-    }
-
-    .dropdown-menu {
-      position: absolute;
-      top: 60px;
-      right: 0;
-      background-color: #444;
-      border-radius: 8px;
-      box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
-      z-index: 1;
-      padding: 10px;
-      transition: opacity 0.3s ease, visibility 0.3s ease;
-      opacity: 1;
-    }
-
-    .dropdown-menu button {
-      background-color: #ff4d4f;
-      color: white;
-      border: none;
-      padding: 10px 20px;
-      border-radius: 5px;
-      cursor: pointer;
-      width: 100%;
-      text-align: left;
-      /* Aligns the text inside the button */
-    }
-
-    .dropdown-menu button:hover {
-      background-color: #ff7875;
-    }
-
-    .dropdown-menu::before {
-      content: '';
-      position: absolute;
-      top: -10px;
-      right: 20px;
-      border-width: 0 8px 8px 8px;
-      border-style: solid;
-      border-color: transparent transparent #444 transparent;
-    }
-
-    .App-footer {
-      position: fixed;
-      bottom: 0;
-      width: 100%;
-      background-color: #282c34;
-      color: white;
-      text-align: center;
-      padding: 10px 0;
-  background-color: #ffffff;
-
+.App-footer {
+  position: fixed;
+  bottom: 0;
+  width: 100%;
+  background-color: white;
+    color: black;
+    text-align: center;
+    padding: 10px 0;
 }
 
 .App-link {
@@ -102,11 +35,11 @@
 }
 
 @keyframes App-logo-spin {
-    from {
-        transform: rotate(0deg);
-      }
+  from {
+      transform: rotate(0deg);
+    }
 
-            to {
-              transform: rotate(360deg);
-            }
+        to {
+          transform: rotate(360deg);
+        }
 }
